fixed psr-0 autoloading and moved to psr-4, also formatted code to be more in line with psr-2
