+++ composer.json
@@ -14,6 +14,6 @@
         "php": ">=5.3.0"
     },
     "autoload": {
-        "psr-0": { "Cassandra": "lib" }
+        "psr-4": { "Cassandra\\Cql\\": "lib/Cassandra/Cql/" }
     }
 }
